pasp
pasp

+++ src/model/CNN.cpp
@@ -99,7 +99,7 @@ double CNN::loss(const std::vector<double>& pred, const std::vector<double>& lab
 }
 
 std::vector<double> CNN::gradient(const std::vector<double>& x, const std::vector<double>& w, 
-		const std::vector<double>& y, std::vector<int>* z) const
+		const std::vector<double>& y, std::vector<double>* z) const
 {
 	// forward
 	vector<vector<vector<double>>> mid; // layer -> feature -> value
