+++ src/model/Kernel.h
@@ -19,7 +19,7 @@ public:
 	virtual double loss(const std::vector<double>& pred, const std::vector<double>& label) const = 0;
 	
 	virtual std::vector<double> gradient(const std::vector<double>& x, const std::vector<double>& w, 
-		const std::vector<double>& y, std::vector<int>* z = nullptr) const = 0;
+		const std::vector<double>& y, std::vector<double>* z = nullptr) const = 0;
 
 	virtual void updateLocalZ(std::vector<double>& zz);
 	virtual std::vector<double> computeDelta(std::vector<double>& zz);
