+++ src/model/LogisticRegression.h
@@ -19,5 +19,5 @@ public:
 	// std::vector<double> gradient(
 	// 	const std::vector<double>& x, const std::vector<double>& w, const std::vector<double>& y) const;
 	std::vector<double> gradient(const std::vector<double>& x, const std::vector<double>& w, 
-		const std::vector<double>& y, std::vector<int>* z = nullptr) const;
+		const std::vector<double>& y, std::vector<double>* z = nullptr) const;
 };
