+++ src/model/MLP.h
@@ -22,7 +22,7 @@ public:
 	// std::vector<double> gradient(
 	// 	const std::vector<double>& x, const std::vector<double>& w, const std::vector<double>& y) const;
 	std::vector<double> gradient(const std::vector<double>& x, const std::vector<double>& w, 
-		const std::vector<double>& y, std::vector<int>* z = nullptr) const;
+		const std::vector<double>& y, std::vector<double>* z = nullptr) const;
 
 private:
 	double getWeight(const std::vector<double>& w, const int layer, const int from, const int to) const;
